using request ID from client as dataplane id flow id

+++ runtime/src/include/data_plane_profiling.h
@@ -32,7 +32,7 @@ static const char *enum_dataplane_op_name[] = {
 
 struct dataplane_profile_info {
     /** Unique id to be assigned to each packet at entry point of dedos */
-    int dp_id;
+    unsigned long int dp_id;
     /** Current count of dataplane log entries (dp_entry_count < MAX_DATAPLANE_ITEMS) */
     int dp_entry_count;
     /** Hop count on logical flow for this item inside dedos system */
@@ -111,7 +111,7 @@ static void log_dp_event(int msu_id, enum_dataplane_op_id dataplane_op_id,
         }
         snprintf(dp_profile_info->dp_log_entries[dp_profile_info->dp_entry_count],
                 MAX_DATAPLANE_LOG_ENTRY_LEN,
-                "%ld, %lu, %d, %d, %d, %d, %s",
+                "%ld, %lu, %lu, %d, %d, %d, %s",
                 ts, self_tid, dp_profile_info->dp_id,dp_profile_info->dp_seq_count,
                 dp_profile_info->dp_entry_count,
                     msu_id, enum_dataplane_op_name[dataplane_op_id]);
